Add an interface for cached values, expressing size explicitly.
The original implementation of the cache code assumed slices of bytes.  I then
converted it to use arbitrary values, and made the cache capacity count entries
rather than size.  Neither of these was especially satisfying, so this change
adds value.Interface, which expresses the ability of a value to report its own
size.

To get a cache with a capacity in entries, the Size method can report 1 for all
values.  To get a cache with a capacity in bytes, the Size method can report a
size in bytes.  Other metrics can be chosen at will, save that Size must never
return a negative value (or the cache operations will panic).